removed AbstractSyncedFlushTests
